# backend/api/static/swagger.yml
openapi: 3.0.0
info:
  title: API Token Management
  description: API for managing application tokens
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
paths:
  /form-containers:
    post:
      summary: Create a new form container
      tags:
        - Form Containers
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormContainerCreate"
      responses:
        "201":
          description: Form container created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  container_id:
                    type: integer
                    example: 1
                  form_id:
                    type: integer
                    example: 10
                  access_token:
                    type: string
                    example: "eyJhbGciOiJIUzI1..."
        "400":
          description: Invalid request data
        "401":
          description: Unauthorized

    get:
      summary: Retrieve all form containers
      tags:
        - Form Containers
      security:
        - BearerAuth: []
      parameters:
        - name: app_ids
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated list of application IDs
        - name: filter
          in: query
          schema:
            type: string
          description: Filter type (e.g., "status")
        - name: status
          in: query
          schema:
            type: string
          description: Status of the form containers (e.g., "open", "reminder", "escalate")
        - name: sort
          in: query
          schema:
            type: string
            default: "desc"
          description: "Sorting order (default: desc)"
      responses:
        "200":
          description: List of form containers
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 15
                  form_containers:
                    type: array
                    items:
                      $ref: "#/components/schemas/FormContainer"
        "401":
          description: Unauthorized
  /form-containers/{access_token}:
    get:
      summary: Retrieve form container details by access token
      tags:
        - Form Containers
      security:
        - BearerAuth: []
      parameters:
        - name: access_token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Form container details retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormContainer"
        "401":
          description: Unauthorized
        "404":
          description: Form container not found
  /form-containers/{container_id}/forms/{form_id}/validate:
    post:
      summary: Validate a form container
      tags:
        - Form Containers
      security:
        - BearerAuth: []
      parameters:
        - name: container_id
          in: path
          required: true
          schema:
            type: integer
        - name: form_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                archive:
                  type: boolean
                  example: true
      responses:
        "200":
          description: Form successfully validated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Form successfully validated."
                  archive_at:
                    type: string
                    format: date-time
                    example: "2025-03-01T12:00:00Z"
        "401":
          description: Unauthorized
        "400":
          description: Form container already validated
  /form-containers/{container_id}/forms/{form_id}/cancel:
    post:
      summary: Cancel a form
      tags:
        - Form Containers
      security:
        - BearerAuth: []
      parameters:
        - name: container_id
          in: path
          required: true
          schema:
            type: integer
        - name: form_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  example: "User no longer interested."
      responses:
        "200":
          description: Form successfully canceled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Form canceled successfully."
                  form_id:
                    type: integer
                    example: 10
                  comment:
                    type: string
                    example: "User no longer interested."
        "401":
          description: Unauthorized
        "400":
          description: Form cannot be canceled
  /forms/{form_id}:
    get:
      summary: Retrieve a form by its ID
      tags:
        - Forms
      security:
        - BearerAuth: []
      parameters:
        - name: app_id
          in: query
          required: true
          schema:
            type: string
          description: Application ID
        - name: form_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Form details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormResponse"
        "401":
          description: Unauthorized
        "404":
          description: Form not found
  /forms/{form_id}/submit-response:
    post:
      summary: Submit a response to a form
      tags:
        - Forms
      security:
        - BearerAuth: []
      parameters:
        - name: form_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormSubmissionRequest"
      responses:
        "200":
          description: Response submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Response submitted successfully"
        "401":
          description: Unauthorized or Form already answered
        "404":
          description: Form or Question not found
...
components:
  schemas:
    FormContainerCreate:
      type: object
      properties:
        title:
          type: string
          example: "Feedback Form"
        description:
          type: string
          example: "A survey form to collect feedback"
        user_email:
          type: string
          format: email
          example: "user@example.com"
        escalate:
          type: boolean
          example: false
        reminder_delay:
          type: integer
          example: 5
        app_id:
          type: string
          example: "700679c8-c988-4335-a0bf-d2a54e9f2e36"
        campaign_id:
          type: integer
          example: 2
        forms:
          type: array
          items:
            type: object
            properties:
              questions:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    FormContainer:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Feedback Form"
        description:
          type: string
          example: "A survey form to collect feedback"
        user_email:
          type: string
          format: email
          example: "user@example.com"
        escalade_email:
          type: string
          format: email
          example: "admin@example.com"
        escalate:
          type: boolean
          example: false
        validated:
          type: boolean
          example: false
        reminder_delay:
          type: integer
          example: 5
        app_id:
          type: string
          example: "700679c8-c988-4335-a0bf-d2a54e9f2e36"
        campaign_id:
          type: integer
          example: 2
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T12:00:00Z"
...
    FormResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Customer Feedback Form"
        description:
          type: string
          example: "A survey form to collect customer feedback."
        status:
          type: string
          example: "open"
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T12:00:00Z"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
    FormSubmissionRequest:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOi..."
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
    Question:
      type: object
      properties:
        id:
          type: integer
          example: 1
        label:
          type: string
          example: "What is your favorite color?"
        type:
          type: string
          example: "text"
        options:
          type: array
          items:
            type: string
          example: ["Red", "Blue", "Green"]
    Answer:
      type: object
      properties:
        id:
          type: integer
          example: 1
        response:
          type: string
          example: "Blue"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT